requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
openai>=1.10.0
anthropic>=0.12.0
google-generativeai>=0.4.0
python-dotenv
requests
